+++ app/src/main/java/com/weather/android/to/subclasses/Coord.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class Coord {
-	 private float lon;
-	 private float lat;
+	 private Double lon;
+	 private Double lat;
 }
