+++ app/src/main/java/com/weather/android/to/subclasses/Main.java
@@ -4,9 +4,9 @@ import lombok.Data;
 
 @Data
 public class Main {
-	 private float temp;
-	 private float pressure;
-	 private float humidity;
-	 private float temp_min;
-	 private float temp_max;
+	 private Double temp;
+	 private Integer pressure;
+	 private Integer humidity;
+	 private Double temp_min;
+	 private Double temp_max;
 }
